refactor: extract local queue protocol from provider
