+++ crates/runner/src/cmd/local/cancel.rs
@@ -10,8 +10,8 @@ use clap::Args;
 
 use crate::error::{RunnerError, RunnerResult};
 use crate::ids::RunId;
+use crate::local_queue;
 use crate::paths::HomePaths;
-use crate::provider::local_queue;
 
 #[derive(Args)]
 pub struct CancelArgs {
