+++ crates/runner/src/local_queue/mod.rs
@@ -1,0 +1,12 @@
+//! Shared local file-queue protocol for `runner local` commands and provider.
+
+mod paths;
+mod state;
+mod types;
+
+pub(crate) use paths::{
+    cancel_path, cancels_dir, claim_path, claims_dir, job_path, jobs_dir, profile_jobs_dir,
+    result_path, results_dir,
+};
+pub(crate) use state::LocalQueue;
+pub(crate) use types::{JobRequest, JobResponse};
