+++ crates/runner/src/provider/local.rs
@@ -8,50 +8,22 @@
 
 use std::collections::{HashMap, HashSet};
 use std::path::PathBuf;
+use std::sync::Arc;
 use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
-use tokio::task::JoinHandle;
 use tokio_util::sync::CancellationToken;
 use tracing::{info, warn};
 
-use super::{ClaimedJob, CompletionAuth, JobCandidate, JobProvider, local_queue};
+use super::local_cancel::{LocalCancelScanner, LocalCancelWatcher};
+use super::{ClaimedJob, CompletionAuth, JobCandidate, JobProvider};
 use crate::ids::RunId;
+use crate::local_queue::{self, JobRequest, JobResponse, LocalQueue};
 use crate::types::{ExecutionContext, HeartbeatState, SandboxReuseResult};
 use sandbox::SandboxId;
 
 /// Poll interval for discovering new job files and local cancel markers.
 const POLL_INTERVAL: Duration = Duration::from_millis(100);
-/// Job request written by `submit` as a `{job_id}.job` file.
-#[derive(serde::Deserialize, serde::Serialize)]
-pub(crate) struct JobRequest {
-    pub(crate) job_id: RunId,
-    pub(crate) prompt: String,
-    pub(crate) cli_agent_type: String,
-    #[serde(default)]
-    pub(crate) vars: Option<HashMap<String, String>>,
-    #[serde(default)]
-    pub(crate) environment: Option<HashMap<String, String>>,
-    #[serde(default)]
-    pub(crate) user_timezone: Option<String>,
-    #[serde(default)]
-    pub(crate) profile: Option<String>,
-    /// Session ID for sandbox reuse across conversation turns.
-    #[serde(default)]
-    pub(crate) session_id: Option<String>,
-    #[serde(default)]
-    pub(crate) feature_flags: Option<HashMap<String, bool>>,
-}
-
-/// Job response written by the runner as a `{job_id}.result` file.
-#[derive(serde::Deserialize, serde::Serialize)]
-pub(crate) struct JobResponse {
-    pub(crate) run_id: RunId,
-    pub(crate) exit_code: i32,
-    pub(crate) error: Option<String>,
-}
-
 /// [`JobProvider`] backed by a file queue in a shared group directory.
 ///
 /// - `discover()` polls supported profile partitions under `jobs/`.
@@ -63,24 +35,12 @@ pub(crate) struct JobResponse {
 /// by capacity or drain mode. `discover()` also performs the same scan as a
 /// fast path, but correctness does not depend on discovery being polled.
 pub struct LocalProvider {
-    group_dir: PathBuf,
+    queue: LocalQueue,
     supported_profiles: Vec<String>,
     profile_cursor: AtomicUsize,
     cancel: CancellationToken,
     cancel_scanner: LocalCancelScanner,
     cancel_watcher: LocalCancelWatcher,
-}
-
-#[derive(Clone)]
-struct LocalCancelScanner {
-    group_dir: PathBuf,
-    cancel_tokens: Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>>,
-    owned_claims: Arc<tokio::sync::Mutex<HashSet<RunId>>>,
-}
-
-struct LocalCancelWatcher {
-    shutdown: CancellationToken,
-    handle: Mutex<Option<JoinHandle<()>>>,
 }
 
 impl LocalProvider {
@@ -108,16 +68,16 @@ impl LocalProvider {
             profiles = ?supported_profiles,
             "local provider watching"
         );
+        let queue = LocalQueue::new(group_dir.clone());
         let owned_claims = Arc::new(tokio::sync::Mutex::new(HashSet::new()));
-        let cancel_scanner =
-            LocalCancelScanner::new(group_dir.clone(), cancel_tokens, owned_claims);
+        let cancel_scanner = LocalCancelScanner::new(queue.clone(), cancel_tokens, owned_claims);
         let cancel_watcher = if start_cancel_watcher {
             LocalCancelWatcher::start(cancel_scanner.clone())
         } else {
             LocalCancelWatcher::disabled()
         };
         Arc::new(Self {
-            group_dir,
+            queue,
             supported_profiles,
             profile_cursor: AtomicUsize::new(0),
             cancel,
@@ -135,282 +95,11 @@ impl LocalProvider {
     ) -> Arc<Self> {
         Self::new_inner(group_dir, supported_profiles, cancel, cancel_tokens, false)
     }
-
-    #[cfg(test)]
-    fn new_with_cancel_watcher(
-        group_dir: PathBuf,
-        supported_profiles: Vec<String>,
-        cancel: CancellationToken,
-        cancel_tokens: Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>>,
-    ) -> Arc<Self> {
-        Self::new_inner(group_dir, supported_profiles, cancel, cancel_tokens, true)
-    }
-}
-
-impl LocalCancelScanner {
-    fn new(
-        group_dir: PathBuf,
-        cancel_tokens: Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>>,
-        owned_claims: Arc<tokio::sync::Mutex<HashSet<RunId>>>,
-    ) -> Self {
-        Self {
-            group_dir,
-            cancel_tokens,
-            owned_claims,
-        }
-    }
-
-    /// Scan for `.cancel` files and trigger the corresponding cancel tokens.
-    ///
-    /// Active markers are deleted only when this runner owns the claim. A token
-    /// can exist before `claim()` succeeds, so ownership is tracked separately
-    /// to avoid stealing another runner's cancel marker. Markers without a
-    /// token are kept while a claim/job may still exist, and are deleted only
-    /// after they no longer have a pending target.
-    async fn scan_cancel_files(&self) {
-        let cancel_ids = self.collect_cancel_ids();
-        if cancel_ids.is_empty() {
-            return;
-        }
-
-        let tokens = self.snapshot_cancel_tokens(&cancel_ids).await;
-        let owned_claims = self.snapshot_owned_claims(&cancel_ids).await;
-
-        for run_id in cancel_ids {
-            if let Some(token) = tokens.get(&run_id) {
-                let was_cancelled = token.is_cancelled();
-                token.cancel();
-                if !was_cancelled {
-                    info!(run_id = %run_id, "local: cancel file detected, cancelling job");
-                }
-                let should_delete =
-                    owned_claims.contains(&run_id) || !self.cancel_has_pending_target(run_id);
-                if should_delete {
-                    let _ = std::fs::remove_file(local_queue::cancel_path(&self.group_dir, run_id));
-                }
-            } else if !self.cancel_has_pending_target(run_id) {
-                let _ = std::fs::remove_file(local_queue::cancel_path(&self.group_dir, run_id));
-            }
-        }
-    }
-
-    fn collect_cancel_ids(&self) -> Vec<RunId> {
-        let cancel_dir = local_queue::cancels_dir(&self.group_dir);
-        let entries = match std::fs::read_dir(&cancel_dir) {
-            Ok(e) => e,
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => return Vec::new(),
-            Err(e) => {
-                warn!(path = %cancel_dir.display(), error = %e, "local: cannot read cancel dir");
-                return Vec::new();
-            }
-        };
-        let mut cancel_ids = Vec::new();
-        let mut seen = HashSet::new();
-        for entry in entries.filter_map(Result::ok) {
-            let path = entry.path();
-            if path.extension().and_then(|e| e.to_str()) != Some("cancel") {
-                continue;
-            }
-            let Some(stem) = path.file_stem().and_then(|s| s.to_str()) else {
-                continue;
-            };
-            let Ok(run_id) = stem.parse::<RunId>() else {
-                continue;
-            };
-            if seen.insert(run_id) {
-                cancel_ids.push(run_id);
-            }
-        }
-        cancel_ids
-    }
-
-    async fn snapshot_cancel_tokens(
-        &self,
-        cancel_ids: &[RunId],
-    ) -> HashMap<RunId, CancellationToken> {
-        let tokens = self.cancel_tokens.lock().await;
-        cancel_ids
-            .iter()
-            .filter_map(|run_id| tokens.get(run_id).cloned().map(|token| (*run_id, token)))
-            .collect()
-    }
-
-    async fn snapshot_owned_claims(&self, cancel_ids: &[RunId]) -> HashSet<RunId> {
-        let owned = self.owned_claims.lock().await;
-        cancel_ids
-            .iter()
-            .copied()
-            .filter(|run_id| owned.contains(run_id))
-            .collect()
-    }
-
-    async fn mark_owned_claim(&self, run_id: RunId) {
-        self.owned_claims.lock().await.insert(run_id);
-    }
-
-    async fn remove_owned_claim(&self, run_id: RunId) {
-        self.owned_claims.lock().await.remove(&run_id);
-    }
-
-    async fn prune_owned_claims_without_tokens(&self) {
-        let owned_ids: Vec<RunId> = {
-            let owned = self.owned_claims.lock().await;
-            if owned.is_empty() {
-                return;
-            }
-            owned.iter().copied().collect()
-        };
-        let stale_ids: Vec<RunId> = {
-            let tokens = self.cancel_tokens.lock().await;
-            owned_ids
-                .into_iter()
-                .filter(|run_id| !tokens.contains_key(run_id))
-                .collect()
-        };
-        if stale_ids.is_empty() {
-            return;
-        }
-
-        let mut owned = self.owned_claims.lock().await;
-        for run_id in stale_ids {
-            owned.remove(&run_id);
-        }
-    }
-
-    fn cancel_has_pending_target(&self, run_id: RunId) -> bool {
-        if self.result_file_has_content(run_id) {
-            return false;
-        }
-        if local_queue::claim_path(&self.group_dir, run_id).exists() {
-            return true;
-        }
-        self.job_file_exists(run_id).unwrap_or(true)
-    }
-
-    fn result_file_has_content(&self, run_id: RunId) -> bool {
-        let result_path = local_queue::result_path(&self.group_dir, run_id);
-        std::fs::metadata(result_path)
-            .map(|metadata| metadata.is_file() && metadata.len() > 0)
-            .unwrap_or(false)
-    }
-
-    fn job_file_exists(&self, run_id: RunId) -> Option<bool> {
-        self.find_job_file(run_id).map(|path| path.is_some())
-    }
-
-    fn find_job_file(&self, run_id: RunId) -> Option<Option<PathBuf>> {
-        let jobs_dir = local_queue::jobs_dir(&self.group_dir);
-        let orgs = match std::fs::read_dir(&jobs_dir) {
-            Ok(entries) => entries,
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => return Some(None),
-            Err(e) => {
-                warn!(path = %jobs_dir.display(), error = %e, "local: cannot scan jobs dir for job file");
-                return None;
-            }
-        };
-
-        for org in orgs.filter_map(Result::ok) {
-            if !org.file_type().map(|ty| ty.is_dir()).unwrap_or(false) {
-                continue;
-            }
-            let org_path = org.path();
-            let profiles = match std::fs::read_dir(&org_path) {
-                Ok(entries) => entries,
-                Err(e) if e.kind() == std::io::ErrorKind::NotFound => continue,
-                Err(e) => {
-                    warn!(path = %org_path.display(), error = %e, "local: cannot scan profile org dir for job file");
-                    return None;
-                }
-            };
-            for profile in profiles.filter_map(Result::ok) {
-                if !profile.file_type().map(|ty| ty.is_dir()).unwrap_or(false) {
-                    continue;
-                }
-                let path = profile.path().join(format!("{run_id}.job"));
-                match std::fs::metadata(&path) {
-                    Ok(_) => return Some(Some(path)),
-                    Err(e) if e.kind() == std::io::ErrorKind::NotFound => {}
-                    Err(e) => {
-                        warn!(run_id = %run_id, path = %path.display(), error = %e, "local: cannot stat job file");
-                        return None;
-                    }
-                }
-            }
-        }
-
-        Some(None)
-    }
-}
-
-impl LocalCancelWatcher {
-    fn start(scanner: LocalCancelScanner) -> Self {
-        let shutdown = CancellationToken::new();
-        let task_shutdown = shutdown.clone();
-        let handle = match tokio::runtime::Handle::try_current() {
-            Ok(handle) => Some(handle.spawn(async move {
-                loop {
-                    scanner.prune_owned_claims_without_tokens().await;
-                    scanner.scan_cancel_files().await;
-                    tokio::select! {
-                        () = task_shutdown.cancelled() => break,
-                        () = tokio::time::sleep(POLL_INTERVAL) => {}
-                    }
-                }
-            })),
-            Err(e) => {
-                warn!(error = %e, "local: cancel watcher not started because no tokio runtime is active");
-                None
-            }
-        };
-
-        Self {
-            shutdown,
-            handle: Mutex::new(handle),
-        }
-    }
-
-    fn disabled() -> Self {
-        let shutdown = CancellationToken::new();
-        shutdown.cancel();
-        Self {
-            shutdown,
-            handle: Mutex::new(None),
-        }
-    }
-
-    async fn shutdown(&self) {
-        self.shutdown.cancel();
-        let handle = self
-            .handle
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner())
-            .take();
-        let Some(handle) = handle else {
-            return;
-        };
-        if let Err(e) = handle.await {
-            warn!(error = %e, "local: cancel watcher task failed");
-        }
-    }
-}
-
-impl Drop for LocalCancelWatcher {
-    fn drop(&mut self) {
-        self.shutdown.cancel();
-        let handle = self
-            .handle
-            .get_mut()
-            .unwrap_or_else(|poisoned| poisoned.into_inner())
-            .take();
-        if let Some(handle) = handle {
-            handle.abort();
-        }
-    }
 }
 
 impl LocalProvider {
     fn find_job_file(&self, run_id: RunId) -> Option<Option<PathBuf>> {
-        self.cancel_scanner.find_job_file(run_id)
+        self.queue.find_job_file(run_id)
     }
 
     fn remove_job_file_if_present(&self, run_id: RunId) -> bool {
@@ -445,7 +134,7 @@ impl LocalProvider {
             else {
                 continue;
             };
-            let profile_dir = match local_queue::profile_jobs_dir(&self.group_dir, profile) {
+            let profile_dir = match local_queue::profile_jobs_dir(self.queue.group_dir(), profile) {
                 Ok(dir) => dir,
                 Err(e) => {
                     warn!(profile, error = %e, "local: invalid supported profile");
@@ -475,7 +164,7 @@ impl LocalProvider {
                 let Ok(job_id) = stem.parse::<RunId>() else {
                     continue;
                 };
-                let claim_path = local_queue::claim_path(&self.group_dir, job_id);
+                let claim_path = local_queue::claim_path(self.queue.group_dir(), job_id);
                 if claim_path.exists() {
                     continue;
                 }
@@ -489,7 +178,7 @@ impl LocalProvider {
     }
 
     fn result_file_has_content(&self, run_id: RunId) -> bool {
-        self.cancel_scanner.result_file_has_content(run_id)
+        self.queue.result_file_has_content(run_id)
     }
 
     fn write_result(&self, run_id: RunId, exit_code: i32, error: Option<&str>) -> bool {
@@ -506,7 +195,7 @@ impl LocalProvider {
             }
         };
 
-        let result_dir = local_queue::results_dir(&self.group_dir);
+        let result_dir = local_queue::results_dir(self.queue.group_dir());
         if let Err(e) = std::fs::create_dir_all(&result_dir) {
             warn!(path = %result_dir.display(), error = %e, "local: failed to create result dir");
             return false;
@@ -514,7 +203,7 @@ impl LocalProvider {
 
         // Atomic write: tmp then rename, so submit never reads a partial file.
         let tmp_file = result_dir.join(format!("{run_id}.{}.result.tmp", RunId::new_v4()));
-        let result_file = local_queue::result_path(&self.group_dir, run_id);
+        let result_file = local_queue::result_path(self.queue.group_dir(), run_id);
         if let Err(e) = std::fs::write(&tmp_file, &json) {
             warn!(run_id = %run_id, error = %e, "local: failed to write result file");
             let _ = std::fs::remove_file(&tmp_file);
@@ -575,12 +264,12 @@ impl JobProvider for LocalProvider {
         };
 
         // Atomic claim via O_EXCL — only the first runner to create the file wins.
-        let claim_dir = local_queue::claims_dir(&self.group_dir);
+        let claim_dir = local_queue::claims_dir(self.queue.group_dir());
         if let Err(e) = std::fs::create_dir_all(&claim_dir) {
             warn!(path = %claim_dir.display(), error = %e, "local: failed to create claim dir");
             return None;
         }
-        let claim_file = local_queue::claim_path(&self.group_dir, run_id);
+        let claim_file = local_queue::claim_path(self.queue.group_dir(), run_id);
         if std::fs::OpenOptions::new()
             .write(true)
             .create_new(true)
@@ -737,15 +426,17 @@ impl JobProvider for LocalProvider {
         self.cancel_scanner.remove_owned_claim(run_id).await;
         if !self.write_result(run_id, exit_code, error) {
             if self.remove_job_file_if_present(run_id) {
-                let _ = std::fs::remove_file(local_queue::cancel_path(&self.group_dir, run_id));
-                let _ = std::fs::remove_file(local_queue::claim_path(&self.group_dir, run_id));
+                let _ =
+                    std::fs::remove_file(local_queue::cancel_path(self.queue.group_dir(), run_id));
+                let _ =
+                    std::fs::remove_file(local_queue::claim_path(self.queue.group_dir(), run_id));
             }
             return;
         }
         // Best-effort cleanup of cancel file (may have been written after the
         // last discover() scan but before the job actually finished).
-        let _ = std::fs::remove_file(local_queue::cancel_path(&self.group_dir, run_id));
-        let _ = std::fs::remove_file(local_queue::claim_path(&self.group_dir, run_id));
+        let _ = std::fs::remove_file(local_queue::cancel_path(self.queue.group_dir(), run_id));
+        let _ = std::fs::remove_file(local_queue::claim_path(self.queue.group_dir(), run_id));
     }
 
     async fn heartbeat(&self, _state: &HeartbeatState) {}
@@ -778,19 +469,6 @@ mod tests {
         tokens: Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>>,
     ) -> Arc<LocalProvider> {
         LocalProvider::new_without_cancel_watcher(
-            dir.to_path_buf(),
-            default_profiles(),
-            cancel,
-            tokens,
-        )
-    }
-
-    fn default_provider_with_cancel_watcher(
-        dir: &std::path::Path,
-        cancel: CancellationToken,
-        tokens: Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>>,
-    ) -> Arc<LocalProvider> {
-        LocalProvider::new_with_cancel_watcher(
             dir.to_path_buf(),
             default_profiles(),
             cancel,
@@ -1187,382 +865,6 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn cancel_file_deleted_after_owned_trigger() {
-        let dir = tempfile::tempdir().unwrap();
-        let cancel = CancellationToken::new();
-        let tokens = empty_cancel_tokens();
-        let run_id = RunId::new_v4();
-        let job_token = CancellationToken::new();
-        tokens.lock().await.insert(run_id, job_token.clone());
-
-        let provider = default_provider(dir.path(), cancel, tokens);
-        write_job(dir.path(), run_id, "owned");
-        let candidate = provider.discover().await.unwrap();
-        assert_eq!(candidate.run_id(), run_id);
-        provider.claim(candidate).await.unwrap();
-
-        let cancel_path = local_queue::cancel_path(dir.path(), run_id);
-        std::fs::create_dir_all(cancel_path.parent().unwrap()).unwrap();
-        std::fs::write(&cancel_path, b"").unwrap();
-
-        provider.cancel_scanner.scan_cancel_files().await;
-        assert!(job_token.is_cancelled(), "cancel token should be triggered");
-        assert!(
-            !cancel_path.exists(),
-            "cancel file should be deleted after triggering an owned token"
-        );
-    }
-
-    #[tokio::test]
-    async fn cancel_file_with_preclaim_token_is_kept() {
-        let dir = tempfile::tempdir().unwrap();
-        let cancel = CancellationToken::new();
-        let tokens = empty_cancel_tokens();
-
-        let run_id = RunId::new_v4();
-        let job_token = CancellationToken::new();
-        tokens.lock().await.insert(run_id, job_token.clone());
-
-        let provider = default_provider(dir.path(), cancel, tokens);
-        write_job(dir.path(), run_id, "preclaim");
-
-        let cancel_path = local_queue::cancel_path(dir.path(), run_id);
-        std::fs::create_dir_all(cancel_path.parent().unwrap()).unwrap();
-        std::fs::write(&cancel_path, b"").unwrap();
-
-        provider.cancel_scanner.scan_cancel_files().await;
-
-        assert!(
-            job_token.is_cancelled(),
-            "pre-claim token should be cancelled"
-        );
-        assert!(
-            cancel_path.exists(),
-            "cancel file should stay until this runner owns the claim"
-        );
-    }
-
-    #[tokio::test]
-    async fn cancel_file_with_preclaim_token_and_other_runner_claim_is_kept() {
-        let dir = tempfile::tempdir().unwrap();
-        let cancel = CancellationToken::new();
-        let tokens = empty_cancel_tokens();
-
-        let run_id = RunId::new_v4();
-        let job_token = CancellationToken::new();
-        tokens.lock().await.insert(run_id, job_token.clone());
-
-        let provider = default_provider(dir.path(), cancel, tokens);
-
-        let cancel_path = local_queue::cancel_path(dir.path(), run_id);
-        let claim_path = local_queue::claim_path(dir.path(), run_id);
-        std::fs::create_dir_all(cancel_path.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(claim_path.parent().unwrap()).unwrap();
-        std::fs::write(&cancel_path, b"").unwrap();
-        std::fs::write(&claim_path, b"").unwrap();
-
-        provider.cancel_scanner.scan_cancel_files().await;
-
-        assert!(
-            job_token.is_cancelled(),
-            "pre-claim token should still observe cancellation"
-        );
-        assert!(
-            cancel_path.exists(),
-            "cancel file should stay for the runner that owns the claim"
-        );
-    }
-
-    #[tokio::test]
-    async fn cancel_file_with_token_but_no_pending_target_is_deleted() {
-        let dir = tempfile::tempdir().unwrap();
-        let cancel = CancellationToken::new();
-        let tokens = empty_cancel_tokens();
-
-        let run_id = RunId::new_v4();
-        let job_token = CancellationToken::new();
-        tokens.lock().await.insert(run_id, job_token.clone());
-
-        let provider = default_provider(dir.path(), cancel, tokens);
-
-        let cancel_path = local_queue::cancel_path(dir.path(), run_id);
-        std::fs::create_dir_all(cancel_path.parent().unwrap()).unwrap();
-        std::fs::write(&cancel_path, b"").unwrap();
-
-        provider.cancel_scanner.scan_cancel_files().await;
-
-        assert!(
-            job_token.is_cancelled(),
-            "stale token should still be cancelled"
-        );
-        assert!(
-            !cancel_path.exists(),
-            "cancel file should be deleted when no claim or job target remains"
-        );
-    }
-
-    #[tokio::test]
-    async fn cancel_file_with_token_terminal_result_and_leftover_claim_is_deleted() {
-        let dir = tempfile::tempdir().unwrap();
-        let cancel = CancellationToken::new();
-        let tokens = empty_cancel_tokens();
-
-        let run_id = RunId::new_v4();
-        let job_token = CancellationToken::new();
-        tokens.lock().await.insert(run_id, job_token.clone());
-
-        let provider = default_provider(dir.path(), cancel, tokens);
-
-        let cancel_path = local_queue::cancel_path(dir.path(), run_id);
-        let result_path = local_queue::result_path(dir.path(), run_id);
-        let claim_path = local_queue::claim_path(dir.path(), run_id);
-        std::fs::create_dir_all(cancel_path.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(result_path.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(claim_path.parent().unwrap()).unwrap();
-        std::fs::write(&cancel_path, b"").unwrap();
-        std::fs::write(&result_path, b"terminal").unwrap();
-        std::fs::write(&claim_path, b"").unwrap();
-
-        provider.cancel_scanner.scan_cancel_files().await;
-
-        assert!(
-            job_token.is_cancelled(),
-            "stale token should still observe the cancel"
-        );
-        assert!(
-            !cancel_path.exists(),
-            "terminal result should let stale token markers be deleted"
-        );
-    }
-
-    #[tokio::test]
-    async fn cancel_watcher_triggers_owned_token_without_discover() {
-        let dir = tempfile::tempdir().unwrap();
-        let cancel = CancellationToken::new();
-        let tokens = empty_cancel_tokens();
-        let provider = default_provider(dir.path(), cancel, Arc::clone(&tokens));
-
-        let run_id = RunId::new_v4();
-        let job_token = CancellationToken::new();
-        tokens.lock().await.insert(run_id, job_token.clone());
-
-        write_job(dir.path(), run_id, "owned");
-        let candidate = provider.discover().await.unwrap();
-        assert_eq!(candidate.run_id(), run_id);
-        provider.claim(candidate).await.unwrap();
-
-        let cancel_path = local_queue::cancel_path(dir.path(), run_id);
-        std::fs::create_dir_all(cancel_path.parent().unwrap()).unwrap();
-        std::fs::write(&cancel_path, b"").unwrap();
-
-        let watcher = LocalCancelWatcher::start(provider.cancel_scanner.clone());
-        tokio::time::timeout(Duration::from_secs(2), job_token.cancelled())
-            .await
-            .expect("cancel watcher should trigger token");
-
-        watcher.shutdown().await;
-    }
-
-    #[tokio::test]
-    async fn cancel_watcher_shutdown_is_idempotent() {
-        let dir = tempfile::tempdir().unwrap();
-        let provider = default_provider_with_cancel_watcher(
-            dir.path(),
-            CancellationToken::new(),
-            empty_cancel_tokens(),
-        );
-
-        tokio::time::timeout(Duration::from_secs(2), provider.shutdown())
-            .await
-            .expect("first shutdown should complete");
-        tokio::time::timeout(Duration::from_secs(2), provider.shutdown())
-            .await
-            .expect("second shutdown should complete");
-    }
-
-    #[tokio::test]
-    async fn owned_claim_without_cancel_token_is_pruned() {
-        let dir = tempfile::tempdir().unwrap();
-        let cancel = CancellationToken::new();
-        let tokens = empty_cancel_tokens();
-        let run_id = RunId::new_v4();
-        tokens.lock().await.insert(run_id, CancellationToken::new());
-
-        let provider = default_provider(dir.path(), cancel, Arc::clone(&tokens));
-        write_job(dir.path(), run_id, "owned");
-        let candidate = provider.discover().await.unwrap();
-        assert_eq!(candidate.run_id(), run_id);
-        provider.claim(candidate).await.unwrap();
-        assert!(
-            provider
-                .cancel_scanner
-                .snapshot_owned_claims(&[run_id])
-                .await
-                .contains(&run_id),
-            "claim should be tracked as locally owned"
-        );
-
-        provider
-            .cancel_scanner
-            .prune_owned_claims_without_tokens()
-            .await;
-        assert!(
-            provider
-                .cancel_scanner
-                .snapshot_owned_claims(&[run_id])
-                .await
-                .contains(&run_id),
-            "owned claim should stay while its cancel token is still active"
-        );
-
-        tokens.lock().await.remove(&run_id);
-        provider
-            .cancel_scanner
-            .prune_owned_claims_without_tokens()
-            .await;
-
-        assert!(
-            !provider
-                .cancel_scanner
-                .snapshot_owned_claims(&[run_id])
-                .await
-                .contains(&run_id),
-            "owned claim should be pruned after token cleanup"
-        );
-    }
-
-    #[tokio::test]
-    async fn cancel_file_without_pending_target_is_deleted() {
-        let dir = tempfile::tempdir().unwrap();
-        let cancel = CancellationToken::new();
-        let tokens = empty_cancel_tokens();
-        let provider = default_provider(dir.path(), cancel, tokens);
-
-        // Write cancel file for a run_id that has no token, claim, or job.
-        let unknown_id = RunId::new_v4();
-        let cancel_path = local_queue::cancel_path(dir.path(), unknown_id);
-        std::fs::create_dir_all(cancel_path.parent().unwrap()).unwrap();
-        std::fs::write(&cancel_path, b"").unwrap();
-
-        let job_id = RunId::new_v4();
-        write_job(dir.path(), job_id, "still works");
-
-        // Should not panic. The stale cancel file is cleaned up.
-        let candidate = provider.discover().await.unwrap();
-        assert_eq!(candidate.run_id(), job_id);
-        assert!(
-            !cancel_path.exists(),
-            "cancel file should be deleted when no pending target exists"
-        );
-    }
-
-    #[tokio::test]
-    async fn cancel_file_with_claim_owned_by_other_runner_is_kept() {
-        let dir = tempfile::tempdir().unwrap();
-        let cancel = CancellationToken::new();
-        let tokens = empty_cancel_tokens();
-        let provider = default_provider(dir.path(), cancel, tokens);
-
-        let run_id = RunId::new_v4();
-        let cancel_path = local_queue::cancel_path(dir.path(), run_id);
-        let claim_path = local_queue::claim_path(dir.path(), run_id);
-        std::fs::create_dir_all(cancel_path.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(claim_path.parent().unwrap()).unwrap();
-        std::fs::write(&cancel_path, b"").unwrap();
-        std::fs::write(&claim_path, b"").unwrap();
-
-        let job_id = RunId::new_v4();
-        write_job(dir.path(), job_id, "still works");
-
-        let candidate = provider.discover().await.unwrap();
-        assert_eq!(candidate.run_id(), job_id);
-        assert!(
-            cancel_path.exists(),
-            "cancel file should be kept when another runner may own the claim"
-        );
-    }
-
-    #[tokio::test]
-    async fn cancel_file_with_terminal_result_and_leftover_job_is_deleted() {
-        let dir = tempfile::tempdir().unwrap();
-        let cancel = CancellationToken::new();
-        let tokens = empty_cancel_tokens();
-        let provider = default_provider(dir.path(), cancel, tokens);
-
-        let run_id = RunId::new_v4();
-        let cancel_path = local_queue::cancel_path(dir.path(), run_id);
-        let result_path = local_queue::result_path(dir.path(), run_id);
-        std::fs::create_dir_all(cancel_path.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(result_path.parent().unwrap()).unwrap();
-        std::fs::write(&cancel_path, b"").unwrap();
-        std::fs::write(&result_path, b"terminal").unwrap();
-        write_job(dir.path(), run_id, "leftover job");
-
-        let job_id = RunId::new_v4();
-        write_job(dir.path(), job_id, "still works");
-
-        let candidate = provider.discover().await.unwrap();
-        assert_eq!(candidate.run_id(), job_id);
-        assert!(
-            !cancel_path.exists(),
-            "cancel file should be deleted when the result is already terminal"
-        );
-    }
-
-    #[tokio::test]
-    async fn cancel_file_with_terminal_result_and_leftover_claim_is_deleted() {
-        let dir = tempfile::tempdir().unwrap();
-        let cancel = CancellationToken::new();
-        let tokens = empty_cancel_tokens();
-        let provider = default_provider(dir.path(), cancel, tokens);
-
-        let run_id = RunId::new_v4();
-        let cancel_path = local_queue::cancel_path(dir.path(), run_id);
-        let result_path = local_queue::result_path(dir.path(), run_id);
-        let claim_path = local_queue::claim_path(dir.path(), run_id);
-        std::fs::create_dir_all(cancel_path.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(result_path.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(claim_path.parent().unwrap()).unwrap();
-        std::fs::write(&cancel_path, b"").unwrap();
-        std::fs::write(&result_path, b"terminal").unwrap();
-        std::fs::write(&claim_path, b"").unwrap();
-
-        let job_id = RunId::new_v4();
-        write_job(dir.path(), job_id, "still works");
-
-        let candidate = provider.discover().await.unwrap();
-        assert_eq!(candidate.run_id(), job_id);
-        assert!(
-            !cancel_path.exists(),
-            "terminal result should make a leftover claim non-pending"
-        );
-    }
-
-    #[tokio::test]
-    async fn cancel_file_with_empty_result_and_pending_job_is_kept() {
-        let dir = tempfile::tempdir().unwrap();
-        let cancel = CancellationToken::new();
-        let tokens = empty_cancel_tokens();
-        let provider = default_provider(dir.path(), cancel, tokens);
-
-        let run_id = RunId::new_v4();
-        let cancel_path = local_queue::cancel_path(dir.path(), run_id);
-        let result_path = local_queue::result_path(dir.path(), run_id);
-        std::fs::create_dir_all(cancel_path.parent().unwrap()).unwrap();
-        std::fs::create_dir_all(result_path.parent().unwrap()).unwrap();
-        std::fs::write(&cancel_path, b"").unwrap();
-        std::fs::write(&result_path, b"").unwrap();
-        write_job(dir.path(), run_id, "pending job");
-
-        let candidate = provider.discover().await.unwrap();
-        assert_eq!(candidate.run_id(), run_id);
-        assert!(
-            cancel_path.exists(),
-            "empty result file is not terminal, so cancel should stay pending"
-        );
-    }
-
-    #[tokio::test]
     async fn complete_cleans_up_cancel_file() {
         let dir = tempfile::tempdir().unwrap();
         let cancel = CancellationToken::new();
@@ -1626,55 +928,6 @@ mod tests {
         assert!(
             !cancel_path.exists(),
             "cancel file should not be stranded after terminal cleanup"
-        );
-    }
-
-    /// Cancel file written before token is inserted (race between submit
-    /// cancel and runner claim). The file should survive until the token
-    /// appears, then be processed on a subsequent scan.
-    #[tokio::test]
-    async fn cancel_file_before_token_survives_until_token_inserted() {
-        let dir = tempfile::tempdir().unwrap();
-        let cancel = CancellationToken::new();
-        let tokens = empty_cancel_tokens();
-        let provider = default_provider(dir.path(), cancel, Arc::clone(&tokens));
-
-        let run_id = RunId::new_v4();
-        let cancel_path = local_queue::cancel_path(dir.path(), run_id);
-        std::fs::create_dir_all(cancel_path.parent().unwrap()).unwrap();
-        std::fs::write(&cancel_path, b"").unwrap();
-
-        // Write a job so discover returns.
-        write_job(dir.path(), run_id, "will be cancelled");
-
-        // First discover: no token yet → cancel file kept, job returned.
-        let candidate = provider.discover().await.unwrap();
-        assert_eq!(candidate.run_id(), run_id);
-        assert!(
-            cancel_path.exists(),
-            "cancel file should survive (no token yet)"
-        );
-
-        // Simulate main loop inserting token after discover returns.
-        let job_token = CancellationToken::new();
-        tokens.lock().await.insert(run_id, job_token.clone());
-
-        // Claim the job so it's no longer discoverable, then write another
-        // job to let discover() return.
-        provider.claim(candidate).await.unwrap();
-        let other_job = RunId::new_v4();
-        write_job(dir.path(), other_job, "next job");
-
-        // Second discover: token exists now → cancel triggered and file deleted.
-        let candidate2 = provider.discover().await.unwrap();
-        assert_eq!(candidate2.run_id(), other_job);
-        assert!(
-            job_token.is_cancelled(),
-            "token should be cancelled on second scan"
-        );
-        assert!(
-            !cancel_path.exists(),
-            "cancel file should be deleted after trigger"
         );
     }
